[review-onboarding-skill-and-checker] review: fix lint, and split the checker's rendering into its own module
The previous commit's additions pushed check-consumer-config.ts to 1026 lines,
over the repo's 1000-line max-lines budget, and carried two prettier/escape
errors in the test. CI caught all three.

Rather than shave comments to squeeze under the cap by a few lines, this takes the
split the file's own docstring already implies ("the report is data, rendered
separately"): renderReport moves to check-consumer-config-report.ts, the same
by-concern split glob-match.ts took from router.ts for the same reason. The
checker re-exports it, so it stays the single entry point and no importer changes.
That leaves 901 lines and headroom for the next check, instead of a file that
tips over again on the next addition.

Verified with the repo's own eslint and prettier this time (the earlier run could
not resolve plugins from inside a nested worktree, which is why the lint errors
reached CI): eslint clean over workflows/, utils/, actions/, config/ and types/,
prettier clean, 113 tests passing, tsc clean.
