+++ workflows/review/lib/check-consumer-config.test.ts
@@ -242,7 +242,7 @@ describe("the add-reviewer contract", () => {
             "    allowed-team-reviewers:\n      - kore\n",
             "",
         ).replace(
-            "    github-token: \${{ secrets.KHAN_ACTIONS_BOT_TOKEN }}\n",
+            "    github-token: ${{ secrets.KHAN_ACTIONS_BOT_TOKEN }}\n",
             "",
         );
         expect(codes(check(inputs), "warning")).toEqual([
@@ -301,8 +301,8 @@ describe("the compiled lock", () => {
     it("is quiet once the maintenance workflow is marked generated", () => {
         const inputs = validInstall();
         inputs[MAINTENANCE_WORKFLOW_PATH] = "# generated by gh aw\n";
-        inputs[".gitattributes"] +=
-            `${MAINTENANCE_WORKFLOW_PATH} linguist-generated=true merge=ours\n`;
+        const marker = `${MAINTENANCE_WORKFLOW_PATH} linguist-generated=true\n`;
+        inputs[".gitattributes"] += marker;
         expect(codes(check(inputs), "warning")).toEqual([]);
     });
